@@ -11,9 +11,13 @@
     "test": "jest"
   },
   "dependencies": {
-    "@sentry/vue": "^5.29.2",
-    "ant-design-vue": "1.7.8",
-    "chart.js": "^2.9.3",
+    "@ant-design/icons-vue": "^6.1.0",
+    "@kyvg/vue3-notification": "^2.6.1",
+    "@sentry/vue": "7.23.0",
+    "@vitejs/plugin-vue": "3.2.0",
+    "ant-design-vue": "^3.2.15",
+    "chart.js": "4.2.0",
+    "chartjs-adapter-date-fns": "^3.0.0",
     "chartjs-plugin-downsample": "^1.1.0",
     "date-fns": "^2.4.1",
     "devour-client": "^2.0.23",
@@ -21,49 +25,44 @@
     "lodash": "^4.17.21",
     "sass": "^1.56.1",
     "store": "2.0.12",
-    "vite": "3.2.3",
-    "vite-plugin-vue2": "2.0.2",
-    "vue": "^2.7.13",
-    "vue-chartjs": "^3.5.0",
-    "vue-i18n": "8.21.0",
-    "vue-notification": "^1.3.16",
-    "vue-router": "^3.0.6",
-    "vue2-daterange-picker": "^0.6.7",
-    "vuex": "^3.1.1",
-    "vuex-router-sync": "^5.0.0"
+    "vite": "3.2.4",
+    "vue": "3.2.45",
+    "vue-chartjs": "5.2.0",
+    "vue-i18n": "9.2.2",
+    "vue-router": "4.1.6",
+    "vue3-gabe-daterange-picker": "^0.1.2",
+    "vuex": "4.1.0",
+    "vuex-router-sync": "^6.0.0-rc.1"
   },
   "devDependencies": {
-    "@babel/core": "7.4.5",
+    "@babel/core": "^7.20.2",
+    "@babel/eslint-parser": "^7.19.1",
     "@babel/plugin-proposal-class-properties": "7.5.5",
     "@babel/plugin-transform-runtime": "7.9.0",
     "@babel/preset-env": "7.9.0",
     "@babel/preset-react": "7.10.4",
-    "@vitejs/plugin-vue2-jsx": "1.0.3",
-    "@vue/cli-plugin-babel": "4.4.3",
-    "@vue/cli-plugin-eslint": "4.4.3",
-    "@vue/cli-service": "4.5.8",
-    "@vue/test-utils": "1.0.2",
+    "@vitejs/plugin-vue-jsx": "2.1.1",
+    "@vue/compiler-sfc": "3.2.45",
+    "@vue/test-utils": "2.2.4",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "10.0.3",
-    "babel-jest": "^24.8.0",
-    "babel-plugin-syntax-jsx": "6.18.0",
-    "babel-plugin-transform-vue-jsx": "3.7.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "14.0.0",
+    "babel-jest": "^27.5.0",
+    "babel-plugin-syntax-jsx": "^6.18.0",
+    "babel-plugin-transform-vue-jsx": "^3.7.0",
+    "eslint": "^8.29.0",
+    "eslint-config-airbnb-base": "^15.0.0",
     "eslint-import-resolver-alias": "1.1.2",
-    "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "5.2.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.8.0",
     "flush-promises": "1.0.2",
-    "jest": "^25.2.7",
+    "jest": "^27.5.0",
     "jest-extended": "0.11.5",
     "jest-transform-stub": "2.0.0",
     "sinon": "9.0.3",
-    "vue-eslint-parser": "^6.0.4",
-    "vue-jest": "3.0.7",
-    "vue-template-compiler": "^2.7.14"
+    "vite-plugin-eslint": "^1.8.1",
+    "vue-eslint-parser": "^9.1.0",
+    "vue3-jest": "27.0.0-alpha.1"
   },
   "browserslist": [
     "> 1%",
