@@ -1,20 +1,65 @@
-<script>
-import { mapGetters } from 'vuex';
-import { Line, mixins } from 'vue-chartjs';
-import Downsample from 'chartjs-plugin-downsample';
+<template>
+  <Line
+    :data="chartData"
+    :options="chartOptions"
+    class="chartjs-render-monitor"
+  />
+</template>
 
-const { reactiveProp } = mixins;
+<script>
+import 'chartjs-adapter-date-fns';
+import Downsample from 'chartjs-plugin-downsample';
+import { mapGetters } from 'vuex';
+import { Line } from 'vue-chartjs';
+import {
+  Chart as ChartJS,
+  PointElement,
+  LineElement,
+  CategoryScale,
+  LinearScale,
+  TimeScale,
+  Legend,
+  Tooltip,
+  Title,
+} from 'chart.js';
+
+ChartJS.register(
+  Title,
+  Legend,
+  Tooltip,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  TimeScale,
+  Tooltip,
+);
+
+ChartJS.register({ id: 'downsample', ...Downsample });
 
 export default {
-  extends: Line,
-  mixins: [reactiveProp],
+  name: 'DataChart',
+  components: {
+    // eslint-disable-next-line vue/no-reserved-component-names
+    Line,
+  },
   props: {
-    options: {
+    chartData: {
+      type: Object,
+      required: true,
+    },
+    chartOptions: {
       type: Object,
       default() {
         return {};
       },
     },
+  },
+  data() {
+    return {
+      plugins: [
+      ],
+    };
   },
   computed: {
     ...mapGetters(['navCollapsed']),
@@ -26,10 +71,6 @@ export default {
         this.$data._chart.resize(); // eslint-disable-line no-underscore-dangle
       }, 100);
     },
-  },
-  mounted() {
-    this.addPlugin(Downsample);
-    this.renderChart(this.chartData, this.options);
   },
 };
 </script>
