@@ -5,15 +5,15 @@
     >
       {{ field.title }}
     </p>
-    <a-form-model
+    <a-form
       ref="form"
       :model="form"
       :rules="rules"
-      @submit.prevent="filter"
+      @finish="filter"
     >
-      <a-form-model-item>
+      <a-form-item>
         <a-radio-group
-          v-model="form.modifier"
+          v-model:value="form.modifier"
         >
           <a-radio value="cont">
             {{ $t('message.cont') }}
@@ -28,24 +28,24 @@
             {{ $t('message.eq') }}
           </a-radio>
         </a-radio-group>
-      </a-form-model-item>
-      <a-form-model-item
-        prop="value"
+      </a-form-item>
+      <a-form-item
+        name="value"
       >
         <a-input
-          v-model="form.value"
+          v-model:value="form.value"
           :placeholder="field.title"
         />
-      </a-form-model-item>
-      <a-form-model-item>
+      </a-form-item>
+      <a-form-item>
         <a-button
           type="primary"
           html-type="submit"
         >
           {{ $t('message.applyFilter') }}
         </a-button>
-      </a-form-model-item>
-    </a-form-model>
+      </a-form-item>
+    </a-form>
   </div>
 </template>
 
@@ -63,6 +63,9 @@ export default {
       },
     },
   },
+  emits: [
+    'filterChange',
+  ],
   data() {
     return {
       form: {
@@ -80,17 +83,12 @@ export default {
   },
   methods: {
     filter() {
-      this.$refs.form.validate((valid) => {
-        if (valid) {
-          this.$emit('filterChange', {
-            modifier: this.form.modifier,
-            value: this.form.value,
-            field: this.field,
-          });
-          this.$refs.form.resetFields();
-        }
-        return valid;
+      this.$emit('filterChange', {
+        modifier: this.form.modifier,
+        value: this.form.value,
+        field: this.field,
       });
+      this.$refs.form.resetFields();
     },
   },
 };
