@@ -1,15 +1,12 @@
-import { shallowMount, createLocalVue } from '@vue/test-utils';
-import Vuex from 'vuex';
-import VueI18n from 'vue-i18n';
+import { shallowMount } from '@vue/test-utils';
+import { createStore } from 'vuex';
+import { createI18n } from 'vue-i18n';
 
 import { ACCOUNT_DETAILS, COMMON_TABLE_ENTITY_EXCLUDED_FIELDS } from '@/constants';
 import VerticalListAnt from '@/components/VerticalListAnt/VerticalListAnt.vue';
 import Accounts from '../Accounts.vue';
 
-const localVue = createLocalVue();
-localVue.use(Vuex);
-localVue.use(VueI18n);
-const i18n = new VueI18n({ locale: 'en' });
+const i18n = createI18n({ locale: 'en' });
 
 describe('Accounts page', () => {
   let storeParams;
@@ -41,9 +38,13 @@ describe('Accounts page', () => {
   it('calls getAccountDetails endpoint on created, if active account is set', () => {
     expect.assertions(1);
 
-    const store = new Vuex.Store(storeParams);
+    const store = createStore(storeParams);
 
-    shallowMount(Accounts, { store, localVue, i18n });
+    shallowMount(Accounts, {
+      global: {
+        plugins: [store, i18n],
+      },
+    });
 
     expect(getAccountDetails).toHaveBeenCalled();
   });
@@ -58,17 +59,25 @@ describe('Accounts page', () => {
         activeAccount: () => null,
       },
     };
-    const store = new Vuex.Store(adjustedStoreParams);
+    const store = createStore(adjustedStoreParams);
 
-    shallowMount(Accounts, { store, localVue, i18n });
+    shallowMount(Accounts, {
+      global: {
+        plugins: [store, i18n],
+      },
+    });
     expect(getAccountDetails).toHaveBeenCalledTimes(0);
   });
 
   it('renders VerticalListAnt component with proper dataSource prop', () => {
     expect.assertions(1);
 
-    const store = new Vuex.Store(storeParams);
-    const wrapper = shallowMount(Accounts, { store, localVue, i18n });
+    const store = createStore(storeParams);
+    const wrapper = shallowMount(Accounts, {
+      global: {
+        plugins: [store, i18n],
+      },
+    });
     const numberOfRowsInAccountInfo = Object.keys(ACCOUNT_DETAILS).length
     - COMMON_TABLE_ENTITY_EXCLUDED_FIELDS.length;
 
@@ -78,8 +87,12 @@ describe('Accounts page', () => {
   it('calls getAccountDetails if active account is changed', () => {
     expect.assertions(1);
 
-    const store = new Vuex.Store(storeParams);
-    const component = shallowMount(Accounts, { store, localVue, i18n });
+    const store = createStore(storeParams);
+    const component = shallowMount(Accounts, {
+      global: {
+        plugins: [store, i18n],
+      },
+    });
 
     component.vm.$options.watch.activeAccount.call(component.vm);
 
