@@ -1,15 +1,12 @@
-import { createLocalVue, shallowMount } from '@vue/test-utils';
-import Vuex from 'vuex';
-import VueI18n from 'vue-i18n';
+import { shallowMount } from '@vue/test-utils';
+import { createStore } from 'vuex';
+import { createI18n } from 'vue-i18n';
+import { COMMON_TABLE_ENTITY_EXCLUDED_FIELDS, NETWORK_DETAILS } from '@/constants';
+import VerticalListAnt from '@/components/VerticalListAnt/VerticalListAnt.vue';
 
 import NetworkDetails from '../NetworkDetails.vue';
-import VerticalListAnt from '@/components/VerticalListAnt/VerticalListAnt.vue';
-import { COMMON_TABLE_ENTITY_EXCLUDED_FIELDS, NETWORK_DETAILS } from '@/constants';
 
-const localVue = createLocalVue();
-localVue.use(Vuex);
-localVue.use(VueI18n);
-const i18n = new VueI18n({ locale: 'en', messages: { en: { message: { networkDetails: 'ALBANIA' } } } });
+const i18n = createI18n({ locale: 'en', messages: { en: { message: { networkDetails: 'ALBANIA' } } } });
 
 const $route = {
   params: {
@@ -45,9 +42,12 @@ describe('NetworkDetails page', () => {
   it('calls getNetworkDetails on created', () => {
     expect.assertions(1);
 
-    const store = new Vuex.Store(storeParams);
+    const store = createStore(storeParams);
     shallowMount(NetworkDetails, {
-      store, localVue, i18n, mocks: { $route },
+      global: {
+        plugins: [store, i18n],
+        mocks: { $route },
+      },
     });
 
     expect(getNetworkDetails).toHaveBeenCalledWith(expect.anything(), $route.params.id);
@@ -62,9 +62,12 @@ describe('NetworkDetails page', () => {
         ...storeParams.getters, networkDetails: () => NETWORK_DETAILS,
       },
     };
-    const store = new Vuex.Store(adjustedStoreParams);
+    const store = createStore(adjustedStoreParams);
     const wrapper = shallowMount(NetworkDetails, {
-      store, localVue, i18n, mocks: { $route },
+      global: {
+        plugins: [store, i18n],
+        mocks: { $route },
+      },
     });
     const numberOfRowsInNetworkInfo = Object.keys(NETWORK_DETAILS).length
     - COMMON_TABLE_ENTITY_EXCLUDED_FIELDS.length - 1; // one additional key for id
