@@ -20,7 +20,7 @@ import { TABLE_HEADERS_ANT } from './constants';
 import locale from './locale';
 
 export default {
-  name: 'Rates',
+  name: 'RatesPage',
   components: {
     DataTableAnt,
   },
