@@ -1,20 +1,17 @@
-import { shallowMount, createLocalVue } from '@vue/test-utils';
-import Vuex from 'vuex';
-import VueI18n from 'vue-i18n';
+import { shallowMount } from '@vue/test-utils';
+import { createStore } from 'vuex';
+import { createI18n } from 'vue-i18n';
 
 import Rates from '../Rates.vue';
 
-const localVue = createLocalVue();
-localVue.use(Vuex);
-localVue.use(VueI18n);
-const i18n = new VueI18n({ locale: 'en' });
+const i18n = createI18n({ locale: 'en' });
 
 describe('Rates page', () => {
   it('calls getRates endpoint on created, if active account is set', () => {
     expect.assertions(1);
 
     const getRates = jest.fn();
-    const store = new Vuex.Store({
+    const store = createStore({
       getters: {
         activeAccount: () => ({ id: 'someId' }),
         rates: () => ({ items: [] }),
@@ -29,9 +26,9 @@ describe('Rates page', () => {
     });
 
     shallowMount(Rates, {
-      store,
-      localVue,
-      i18n,
+      global: {
+        plugins: [store, i18n],
+      },
     });
     expect(getRates).toHaveBeenCalled();
   });
