@@ -3,7 +3,7 @@ import { NETWORK_SERVICE } from '@/constants';
 export const wrapWithAsyncRequestStatus = async (commit, callback) => {
   commit(NETWORK_SERVICE.MUTATIONS.SWITCH_PENDING_STATE, true, { root: true });
   // eslint-disable-next-line no-console
-  await callback().catch((e) => { console.error(e[0]?.title); });
+  await callback().catch((e) => { console.warn(e[0]?.title); });
   commit(NETWORK_SERVICE.MUTATIONS.SWITCH_PENDING_STATE, false, { root: true });
 };
 
